feat: ReservationRepository를 인터페이스로 변경, SimpleReservationRepository 구현
