+++ src/main/java/roomescape/domain/reservation/model/Reservation.java
@@ -2,22 +2,22 @@ package roomescape.domain.reservation.model;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.concurrent.atomic.AtomicLong;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
 public class Reservation {
-    private static final AtomicLong INDEX = new AtomicLong(0);
 
+    @Setter
     private Long id;
     private String name;
     private LocalDate date;
     private LocalTime time;
 
     @Builder
-    private Reservation(String name, LocalDate date, LocalTime time) {
-        this.id = INDEX.incrementAndGet();
+    private Reservation(Long id, String name, LocalDate date, LocalTime time) {
+        this.id = id;
         this.name = name;
         this.date = date;
         this.time = time;
